requirements.txt
asgiref==3.3.1
Django==3.1.5
django-crispy-forms==1.10.0
django-taggit==1.3.0
mysqlclient==2.0.3
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
